refactor(tests): unit tests co-located, integration tests moved to tests/integration/
